GEOMESA-3246 Upgrade dependency versions
* Remove Scala 2.11 build

* Apache Arrow 0.16.0 -> 10.0.1
* Apache Orc 1.5.4 -> 1.8.0
* Apache Parquet 1.9.0 -> 1.12.3
* Apache Avro 1.8.2 -> 1.11.1
* Netty 4.1.48.Final -> 4.1.85.Final
* com.fasterxml.jackson -> 2.13.x -> 2.14.1
* Apache Cassandra 3.0.0 -> 3.11.3
* Apache Kafka 2.1.1 -> 2.8.2
* ZKClient 0.10 -> 0.11
* Confluent 5.1.0 -> 6.2.7
* Parboiled 1.2.0 -> 1.3.1
* Apache httpcore 4.4.10 -> 4.4.15
* Spring-security 5.1.13.RELEASE -> 5.8.0
* Commons-csv 1.5 -> 1.9.0
* JSON-path 2.2.0 -> 2.7.0
* JSON-smark 2.2.1 -> 2.4.8
* Aircompressor 0.10 -> 0.21
* Paranamer 2.6 -> 2.8
* Caffeine 2.5.6 -> 2.9.3
* Gson 2.8.9 -> 2.10
* H2 1.4.192 -> 2.1.214
* Chill 0.8.5 -> 0.10.0
* Typesafe Config 1.3.3 -> 1.4.2
* Jedis 3.0.1 -> 4.3.1
* Mango 1.2.0 -> 3.0.0
* Sizeof 0.3.0 -> 0.4.0

+++ geomesa-lambda/geomesa-lambda-datastore/src/test/scala/org/locationtech/geomesa/lambda/stream/kafka/KafkaStoreTest.scala
@@ -45,7 +45,7 @@ class KafkaStoreTest extends LambdaTest with LazyLogging {
     val props = new Properties()
     props.put("bootstrap.servers", brokers)
     WithClose(AdminClient.create(props)) { admin =>
-      admin.createTopics(Collections.singletonList(new NewTopic(topic, 2, 1))).all().get
+      admin.createTopics(Collections.singletonList(new NewTopic(topic, 2, 1.toShort))).all().get
     }
     logger.trace(s"created topic $topic")
   }
